Added common test free joint features (#363)
